+++ demos/turborepo-next-monorepo-demo/apps/web/pages/api/edge.js
@@ -1,0 +1,7 @@
+// Next.js API route support: https://nextjs.org/docs/api-routes/introduction
+
+export default function handler(req, res) {
+  res.status(200).json({
+    body: 'This should not be returned',
+  });
+}