+++ packages/runtime/src/index.ts
@@ -26,7 +26,7 @@ import {
 import { moveStaticPages, movePublicFiles, patchNextFiles, writeStaticRouteManifest } from './helpers/files'
 import { generateFunctions, setupImageFunction, generatePagesResolver } from './helpers/functions'
 import { generateRedirects, generateStaticRedirects } from './helpers/redirects'
-import { shouldSkip, isNextAuthInstalled, getCustomImageResponseHeaders } from './helpers/utils'
+import { shouldSkip, isNextAuthInstalled, getCustomImageResponseHeaders, getRemotePatterns } from './helpers/utils'
 import {
   verifyNetlifyBuildVersion,
   checkNextSiteHasBuilt,
@@ -65,6 +65,7 @@ const plugin: NetlifyPlugin = {
     netlifyConfig.build.environment.NEXT_PRIVATE_TARGET = 'server'
   },
 
+  // eslint-disable-next-line max-lines-per-function
   async onBuild({
     constants,
     netlifyConfig,
@@ -79,14 +80,12 @@ const plugin: NetlifyPlugin = {
 
     checkNextSiteHasBuilt({ publish, failBuild })
 
-    let experimentalRemotePatterns = []
     const { appDir, basePath, i18n, images, target, ignore, trailingSlash, outdir, experimental } = await getNextConfig(
       {
         publish,
         failBuild,
       },
     )
-
     await cleanupEdgeFunctions(constants)
 
     const middlewareManifest = await loadMiddlewareManifest(netlifyConfig)
@@ -117,10 +116,6 @@ const plugin: NetlifyPlugin = {
       }
     }
 
-    if (experimental.images) {
-      experimentalRemotePatterns = experimental.images.remotePatterns || []
-    }
-
     if (isNextAuthInstalled()) {
       const config = await getRequiredServerFiles(publish)
 
@@ -128,10 +123,20 @@ const plugin: NetlifyPlugin = {
 
       if (userDefinedNextAuthUrl) {
         console.log(
-          `NextAuth package detected, NEXTAUTH_URL environment variable set by user to ${userDefinedNextAuthUrl}`,
+          `NextAuth package detected, NEXTAUTH_URL environment variable set by user in next.config.js to ${userDefinedNextAuthUrl}`,
         )
+      } else if (process.env.NEXTAUTH_URL) {
+        // When the value is specified in the netlify.toml or the Netlify UI (will be evaluated in this order)
+        const nextAuthUrl = `${process.env.NEXTAUTH_URL}${basePath}`
+
+        console.log(
+          `NextAuth package detected, NEXTAUTH_URL environment variable set by user in Netlify configuration to ${nextAuthUrl}`,
+        )
+        config.config.env.NEXTAUTH_URL = nextAuthUrl
+
+        await updateRequiredServerFiles(publish, config)
       } else {
-        const nextAuthUrl = `${process.env.URL}${basePath}`
+        const nextAuthUrl = `${process.env.DEPLOY_PRIME_URL}${basePath}`
 
         console.log(`NextAuth package detected, setting NEXTAUTH_URL environment variable to ${nextAuthUrl}`)
         config.config.env.NEXTAUTH_URL = nextAuthUrl
@@ -169,7 +174,7 @@ const plugin: NetlifyPlugin = {
       imageconfig: images,
       netlifyConfig,
       basePath,
-      remotePatterns: experimentalRemotePatterns,
+      remotePatterns: getRemotePatterns(experimental, images),
       responseHeaders: getCustomImageResponseHeaders(netlifyConfig.headers),
     })